moved default configuration
